[test-gen] Write MacOS tests into separate file
