+++ tests/shaping/in_house.rs
@@ -8333,2349 +8333,6 @@ fn ligature_id_038() {
     );
 }
 
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_016() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/BigCaslon.ttf",
-            "\u{0107}",
-            "",
-        ),
-        "cacute=0+432"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_020() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Apple Color Emoji.ttc",
-            "\u{1F1FA}\u{1F1F8}\u{1F1EE}\u{1F1F1}",
-            "--script=hebrew --direction ltr",
-        ),
-        "u1F1FA_u1F1F8=0+800|\
-         u1F1EE_u1F1F1=2+800"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_023() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Bangla MN.ttc",
-            "\u{09AC}\u{09BF}",
-            "",
-        ),
-        "bn_ikaar=0+474|\
-         bn_ba=0+998"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_026() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{20DD}\u{1F174}\u{1F175}",
-            "",
-        ),
-        "circlecmb=0+0|\
-         .notdef=1+1536|\
-         .notdef=2+1536"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_029() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{0041}\u{0042}\u{0043}\u{0044}",
-            "--font-funcs ot --show-flags",
-        ),
-        "A=0+1413|\
-         B=1+1178|\
-         C=2+1417|\
-         D=3+1534"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_032() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{006D}\u{0300}",
-            "--font-funcs ot",
-        ),
-        "mgrave=0+1912"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_033() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Khmer MN.ttc",
-            "\u{17A2}\u{1780}\u{17D2}\u{179F}\u{179A}\u{1781}\u{17D2}\u{1798}\u{17C2}\u{179A}",
-            "--font-funcs ot",
-        ),
-        "km_qa=0+1230|\
-         km_ka=1+1230|\
-         km_sa.sub=1+620|\
-         km_ro=4+712|\
-         km_vs_ae=5+726|\
-         km_kha=5+1230|\
-         km_mo.sub=5+0|\
-         km_ro=9+712"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_034() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Tamil MN.ttc",
-            "\u{0BA4}\u{0BCA}\u{0B95}\u{0BC1}\u{0B95}\u{0BCD}\u{0B95}\u{0BAA}\u{0BCD}\u{0BAA}\
-             \u{0B9F}\u{0BCD}\u{0B9F}\u{0BC1}",
-            "--font-funcs ot",
-        ),
-        "tgm_e=0+1702|\
-         tgc_ta=0+1598|\
-         tgm_aa=0+1149|\
-         tgc_ku=2+1962|\
-         tgc_k=4+1592|\
-         tgc_ka=6+1592|\
-         tgc_p=7+1370|\
-         tgc_pa=9+1370|\
-         tgc_tt=10+1596|\
-         tgc_ttu=12+1833"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_035() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{05E1}\u{05B0}",
-            "--font-funcs ot",
-        ),
-        "shevahebrew=0@51,0+0|\
-         samekhhebrew=0+1361"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_036() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0054}\u{0068}\u{0020}\u{0074}\u{0068}\u{0020}\u{006C}\u{006C}\u{0020}\u{0074}\
-             \u{0065}\u{0020}\u{0074}\u{006F}\u{0020}\u{0074}\u{0072}\u{0020}\u{0066}\u{0072}\
-             \u{0020}\u{0066}\u{0075}\u{0020}\u{0066}\u{006A}",
-            "--font-funcs ot",
-        ),
-        "T_h=0+2308|\
-         space=2+569|\
-         t_h=3+1687|\
-         space=5+569|\
-         l_l=6+1108|\
-         space=8+569|\
-         t_e=9+1408|\
-         space=11+569|\
-         t_o=12+1531|\
-         space=14+569|\
-         t_r=15+1385|\
-         space=17+569|\
-         f_r=18+1432|\
-         space=20+569|\
-         f_u=21+1733|\
-         space=23+569|\
-         f_j=24+1098"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_037() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0054}\u{0065}\u{0020}\u{0041}\u{0056}\u{0020}\u{0054}\u{0072}\u{0020}\u{0056}\
-             \u{0061}\u{0020}\u{0072}\u{0054}\u{0020}\u{0065}\u{0054}\u{0020}\u{0054}\u{0064}",
-            "--font-funcs ot",
-        ),
-        "T=0+1497|\
-         e=1@-62,0+699|\
-         space=2+569|\
-         A=3+1431|\
-         V=4@-37,0+1377|\
-         space=5+569|\
-         T=6+1510|\
-         r=7@-50,0+803|\
-         space=8+569|\
-         V=9+1376|\
-         a=10@-37,0+1014|\
-         space=11+569|\
-         r=12+853|\
-         T=13+1560|\
-         space=14+569|\
-         e=15+761|\
-         T=16+1560|\
-         space=17+569|\
-         T=18+1515|\
-         d=19@-45,0+1006"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_038() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0627}\u{0644}\u{0623}\u{064E}\u{0628}\u{0652}\u{062C}\u{064E}\u{062F}\u{0650}\
-             \u{064A}\u{064E}\u{0651}\u{0629}\u{0640}\u{0627}\u{0644}\u{0639}\u{064E}\u{0631}\
-             \u{064E}\u{0628}\u{0650}\u{064A}\u{064E}\u{0651}\u{0629}",
-            "--font-funcs ot",
-        ),
-        "u0629.final.tehMarbuta=26+713|\
-         u064e_u0651.shaddaFatha=23@0,-200+0|\
-         u064a.medial.yeh=23+656|\
-         u0650.kasra=21@80,290+80|\
-         u0628.initial.beh=21@-80,0+576|\
-         u064e.fatha=19@200,-570+200|\
-         u0631.final.reh=19@-200,0+702|\
-         u064e.fatha=17@200,-200+200|\
-         u0639.medial.ain=17@-200,0+738|\
-         u0644.initial.lam=16+515|\
-         u0627.final.alef=15+647|\
-         u0640.tatweel=14+449|\
-         u0629.final.tehMarbuta=13+713|\
-         u064e_u0651.shaddaFatha=10@0,-200+0|\
-         u064a.initial.yeh=10+656|\
-         u0650.kasra=8@80,570+80|\
-         u062f.final.dal=8@-80,0+822|\
-         u064e.fatha=6@290,-160+290|\
-         u062c.medial.jeem=6@-290,0+1069|\
-         u0652.sukun=4@0,-200+0|\
-         u0628.initial.beh=4+656|\
-         u064e.fatha=1@-372,120+-372|\
-         u0644_u0623.isolated.lamHamzaOnAlef=1@120,0+1282|\
-         u0627.alef=0+647"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_039() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0628}\u{064A}\u{064E}\u{0651}\u{0629}",
-            "--font-funcs ot",
-        ),
-        "u0629.final.tehMarbuta=4+713|\
-         u064e_u0651.shaddaFatha=1@0,-200+0|\
-         u064a.medial.yeh=1+656|\
-         u0628.initial.beh=0+656"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_040() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0631}\u{0628}",
-            "--font-funcs ot",
-        ),
-        "u0628.beh=1+1415|\
-         u0631.reh=0@-202,0+700"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_041() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0628}\u{064F}",
-            "--font-funcs ot",
-        ),
-        "u064f.damma=0@250,-250+250|\
-         u0628.beh=0@-250,0+1165"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_042() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0644}\u{064E}\u{0645}\u{064E}\u{0651}\u{0627}",
-            "--font-funcs ot",
-        ),
-        "u0627.final.alef=5+647|\
-         u064e.fatha=0@-80,160+-80|\
-         u064e_u0651.shaddaFatha=0@490,250+490|\
-         u0644_u0645.initial.lamMeem=0@-410,0+415"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_043() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Apple Color Emoji.ttc",
-            "\u{1F468}\u{200D}\u{1F469}\u{200D}\u{1F467}\u{200D}\u{1F466}",
-            "--remove-default-ignorables --font-funcs ot",
-        ),
-        "u1F9D1_u1F9D1_u1F9D2_u1F9D2=0+800"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_044() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Zapfino.ttf",
-            "\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\
-             \u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\
-             \u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\
-             \u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\
-             \u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{006F}",
-            "--font-funcs ot",
-        ),
-        "Z=0+416|\
-         a=1@-21,0+264|\
-         p_f=2+433|\
-         i=4+181|\
-         n=5+261|\
-         Z=6+416|\
-         a=7@-21,0+264|\
-         p_f=8+433|\
-         i=10+181|\
-         n=11+261|\
-         Z=12+416|\
-         a=13@-21,0+264|\
-         p_f=14+433|\
-         i=16+181|\
-         n=17+261|\
-         Z=18+416|\
-         a=19@-21,0+264|\
-         p_f=20+433|\
-         i=22+181|\
-         n=23+261|\
-         Z=24+416|\
-         a=25@-21,0+264|\
-         p_f=26+433|\
-         i=28+181|\
-         n=29+261|\
-         Z=30+416|\
-         a=31@-21,0+264|\
-         p_f=32+433|\
-         i=34+181|\
-         n=35+261|\
-         Z=36+416|\
-         a=37@-21,0+264|\
-         p_f=38+433|\
-         i=40+181|\
-         n=41+261|\
-         Z=42+416|\
-         a=43@-21,0+264|\
-         p_f=44+433|\
-         i=46+181|\
-         n=47+261|\
-         Z=48+416|\
-         a=49@-21,0+264|\
-         p_f=50+433|\
-         i=52+181|\
-         n=53+261|\
-         Z=54+416|\
-         a=55@-21,0+264|\
-         p_f=56+433|\
-         i=58+181|\
-         n=59+261|\
-         Z=60+416|\
-         a=61@-21,0+264|\
-         p_f=62+433|\
-         i=64+181|\
-         n=65+261|\
-         Z_a_p_f_i_n_o=66+2333"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_046() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Kokonor.ttf",
-            "\u{0F62}\u{0F92}\u{0FB1}\u{0F74}",
-            "--font-funcs ot",
-        ),
-        "r_g_y_u=0+1579"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_047() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Kokonor.ttf",
-            "\u{0F62}\u{0F92}\u{0FB1}\u{0F74}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "r_g_y_u=0+1579"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_048() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_049() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features dlig=1",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_050() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_051() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features dlig=1,sups=0,tnum=0",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_052() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0,sups=0,tnum=0",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_053() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features smcp=1",
-        ),
-        "F.small=0+903|\
-         I.small=1+634|\
-         R.small=2+1113|\
-         S.small=3+911|\
-         T.small=4+1075"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_054() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0,dlig=1",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_057() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Helvetica.ttc",
-            "\u{006D}\u{0300}",
-            "--font-funcs ot",
-        ),
-        "m=0+1706|\
-         gravecmb=0@-284,10+0"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_058() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{006D}\u{0300}",
-            "--font-funcs ot",
-        ),
-        "mgrave=0+1912"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_059() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0066}\u{0069}",
-            "--font-funcs ot",
-        ),
-        "fi=0+1139"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_060() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Khmer MN.ttc",
-            "\u{17A2}\u{1780}\u{17D2}\u{179F}\u{179A}\u{1781}\u{17D2}\u{1798}\u{17C2}\u{179A}",
-            "--font-funcs ot",
-        ),
-        "km_qa=0+1230|\
-         km_ka=1+1230|\
-         km_sa.sub=1+620|\
-         km_ro=4+712|\
-         km_vs_ae=5+726|\
-         km_kha=5+1230|\
-         km_mo.sub=5+0|\
-         km_ro=9+712"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_061() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Tamil MN.ttc",
-            "\u{0BA4}\u{0BCA}\u{0B95}\u{0BC1}\u{0B95}\u{0BCD}\u{0B95}\u{0BAA}\u{0BCD}\u{0BAA}\
-             \u{0B9F}\u{0BCD}\u{0B9F}\u{0BC1}",
-            "--font-funcs ot",
-        ),
-        "tgm_e=0+1702|\
-         tgc_ta=0+1598|\
-         tgm_aa=0+1149|\
-         tgc_ku=2+1962|\
-         tgc_k=4+1592|\
-         tgc_ka=6+1592|\
-         tgc_p=7+1370|\
-         tgc_pa=9+1370|\
-         tgc_tt=10+1596|\
-         tgc_ttu=12+1833"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_062() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0041}\u{0066}\u{0300}\u{0066}\u{0069}\u{005A}",
-            "--font-funcs ot",
-        ),
-        "A=0+1479|\
-         f=1+682|\
-         gravecmb=1@-551,588+0|\
-         fi=3+1139|\
-         Z=5+1251"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_063() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{05E1}\u{05B0}",
-            "--font-funcs ot",
-        ),
-        "shevahebrew=0@51,0+0|\
-         samekhhebrew=0+1361"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_064() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0054}\u{0068}\u{0020}\u{0074}\u{0068}\u{0020}\u{006C}\u{006C}\u{0020}\u{0074}\
-             \u{0065}\u{0020}\u{0074}\u{006F}\u{0020}\u{0074}\u{0072}\u{0020}\u{0066}\u{0072}\
-             \u{0020}\u{0066}\u{0075}\u{0020}\u{0066}\u{006A}",
-            "--font-funcs ot",
-        ),
-        "T_h=0+2308|\
-         space=2+569|\
-         t_h=3+1687|\
-         space=5+569|\
-         l_l=6+1108|\
-         space=8+569|\
-         t_e=9+1408|\
-         space=11+569|\
-         t_o=12+1531|\
-         space=14+569|\
-         t_r=15+1385|\
-         space=17+569|\
-         f_r=18+1432|\
-         space=20+569|\
-         f_u=21+1733|\
-         space=23+569|\
-         f_j=24+1098"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_065() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0054}\u{0065}\u{0020}\u{0041}\u{0056}\u{0020}\u{0054}\u{0072}\u{0020}\u{0056}\
-             \u{0061}\u{0020}\u{0072}\u{0054}\u{0020}\u{0065}\u{0054}\u{0020}\u{0054}\u{0064}",
-            "--font-funcs ot",
-        ),
-        "T=0+1497|\
-         e=1@-62,0+699|\
-         space=2+569|\
-         A=3+1431|\
-         V=4@-37,0+1377|\
-         space=5+569|\
-         T=6+1510|\
-         r=7@-50,0+803|\
-         space=8+569|\
-         V=9+1376|\
-         a=10@-37,0+1014|\
-         space=11+569|\
-         r=12+853|\
-         T=13+1560|\
-         space=14+569|\
-         e=15+761|\
-         T=16+1560|\
-         space=17+569|\
-         T=18+1515|\
-         d=19@-45,0+1006"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_066() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0627}\u{0644}\u{0623}\u{064E}\u{0628}\u{0652}\u{062C}\u{064E}\u{062F}\u{0650}\
-             \u{064A}\u{064E}\u{0651}\u{0629}\u{0640}\u{0627}\u{0644}\u{0639}\u{064E}\u{0631}\
-             \u{064E}\u{0628}\u{0650}\u{064A}\u{064E}\u{0651}\u{0629}",
-            "--font-funcs ot",
-        ),
-        "u0629.final.tehMarbuta=26+713|\
-         u064e_u0651.shaddaFatha=23@0,-200+0|\
-         u064a.medial.yeh=23+656|\
-         u0650.kasra=21@80,290+80|\
-         u0628.initial.beh=21@-80,0+576|\
-         u064e.fatha=19@200,-570+200|\
-         u0631.final.reh=19@-200,0+702|\
-         u064e.fatha=17@200,-200+200|\
-         u0639.medial.ain=17@-200,0+738|\
-         u0644.initial.lam=16+515|\
-         u0627.final.alef=15+647|\
-         u0640.tatweel=14+449|\
-         u0629.final.tehMarbuta=13+713|\
-         u064e_u0651.shaddaFatha=10@0,-200+0|\
-         u064a.initial.yeh=10+656|\
-         u0650.kasra=8@80,570+80|\
-         u062f.final.dal=8@-80,0+822|\
-         u064e.fatha=6@290,-160+290|\
-         u062c.medial.jeem=6@-290,0+1069|\
-         u0652.sukun=4@0,-200+0|\
-         u0628.initial.beh=4+656|\
-         u064e.fatha=1@-372,120+-372|\
-         u0644_u0623.isolated.lamHamzaOnAlef=1@120,0+1282|\
-         u0627.alef=0+647"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_067() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0628}\u{064A}\u{064E}\u{0651}\u{0629}",
-            "--font-funcs ot",
-        ),
-        "u0629.final.tehMarbuta=4+713|\
-         u064e_u0651.shaddaFatha=1@0,-200+0|\
-         u064a.medial.yeh=1+656|\
-         u0628.initial.beh=0+656"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_068() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0631}\u{0628}",
-            "--font-funcs ot",
-        ),
-        "u0628.beh=1+1415|\
-         u0631.reh=0@-202,0+700"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_069() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0628}\u{064F}",
-            "--font-funcs ot",
-        ),
-        "u064f.damma=0@250,-250+250|\
-         u0628.beh=0@-250,0+1165"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_070() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0644}\u{064E}\u{0645}\u{064E}\u{0651}\u{0627}",
-            "--font-funcs ot",
-        ),
-        "u0627.final.alef=5+647|\
-         u064e.fatha=0@-80,160+-80|\
-         u064e_u0651.shaddaFatha=0@490,250+490|\
-         u0644_u0645.initial.lamMeem=0@-410,0+415"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_071() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Apple Color Emoji.ttc",
-            "\u{1F468}\u{200D}\u{1F469}\u{200D}\u{1F467}\u{200D}\u{1F466}",
-            "--remove-default-ignorables --font-funcs ot",
-        ),
-        "u1F9D1_u1F9D1_u1F9D2_u1F9D2=0+800"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_072() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Zapfino.ttf",
-            "\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\
-             \u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\
-             \u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\
-             \u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\
-             \u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{006F}",
-            "--font-funcs ot",
-        ),
-        "Z=0+416|\
-         a=1@-21,0+264|\
-         p_f=2+433|\
-         i=4+181|\
-         n=5+261|\
-         Z=6+416|\
-         a=7@-21,0+264|\
-         p_f=8+433|\
-         i=10+181|\
-         n=11+261|\
-         Z=12+416|\
-         a=13@-21,0+264|\
-         p_f=14+433|\
-         i=16+181|\
-         n=17+261|\
-         Z=18+416|\
-         a=19@-21,0+264|\
-         p_f=20+433|\
-         i=22+181|\
-         n=23+261|\
-         Z=24+416|\
-         a=25@-21,0+264|\
-         p_f=26+433|\
-         i=28+181|\
-         n=29+261|\
-         Z=30+416|\
-         a=31@-21,0+264|\
-         p_f=32+433|\
-         i=34+181|\
-         n=35+261|\
-         Z=36+416|\
-         a=37@-21,0+264|\
-         p_f=38+433|\
-         i=40+181|\
-         n=41+261|\
-         Z=42+416|\
-         a=43@-21,0+264|\
-         p_f=44+433|\
-         i=46+181|\
-         n=47+261|\
-         Z=48+416|\
-         a=49@-21,0+264|\
-         p_f=50+433|\
-         i=52+181|\
-         n=53+261|\
-         Z=54+416|\
-         a=55@-21,0+264|\
-         p_f=56+433|\
-         i=58+181|\
-         n=59+261|\
-         Z=60+416|\
-         a=61@-21,0+264|\
-         p_f=62+433|\
-         i=64+181|\
-         n=65+261|\
-         Z_a_p_f_i_n_o=66+2333"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_073() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Thonburi.ttc",
-            "\u{0E17}\u{0E35}\u{0E48}\u{0E4A}",
-            "--font-funcs ot",
-        ),
-        "thothahanthai_saraiithai_maiekthai=0+1616|\
-         maitrithai.key=0+1300"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_074() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0066}\u{0069}\u{006e}\u{0065}",
-            "--font-funcs ot",
-        ),
-        "fi=0+1139|\
-         n=2+1024|\
-         e=3+909"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_075() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0066}\u{0069}\u{006e}\u{0065}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "f=0+682|\
-         i=1+569|\
-         n=2+1024|\
-         e=3+909"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_076() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Kokonor.ttf",
-            "\u{0F62}\u{0F92}\u{0FB1}\u{0F74}",
-            "--font-funcs ot",
-        ),
-        "r_g_y_u=0+1579"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_077() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Kokonor.ttf",
-            "\u{0F62}\u{0F92}\u{0FB1}\u{0F74}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "r_g_y_u=0+1579"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_078() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_079() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features dlig=1",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_080() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_081() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features dlig=1,sups=0,tnum=0",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_082() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0,sups=0,tnum=0",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_083() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features smcp=1",
-        ),
-        "F.small=0+903|\
-         I.small=1+634|\
-         R.small=2+1113|\
-         S.small=3+911|\
-         T.small=4+1075"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_084() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0,dlig=1",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_087() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Helvetica.ttc",
-            "\u{006D}\u{0300}",
-            "--font-funcs ot",
-        ),
-        "m=0+1706|\
-         gravecmb=0@-284,10+0"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_088() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{006D}\u{0300}",
-            "--font-funcs ot",
-        ),
-        "mgrave=0+1912"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_089() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0066}\u{0069}",
-            "--font-funcs ot",
-        ),
-        "fi=0+1139"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_090() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Khmer MN.ttc",
-            "\u{17A2}\u{1780}\u{17D2}\u{179F}\u{179A}\u{1781}\u{17D2}\u{1798}\u{17C2}\u{179A}",
-            "--font-funcs ot",
-        ),
-        "km_qa=0+1230|\
-         km_ka=1+1230|\
-         km_sa.sub=1+620|\
-         km_ro=4+712|\
-         km_vs_ae=5+726|\
-         km_kha=5+1230|\
-         km_mo.sub=5+0|\
-         km_ro=9+712"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_091() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Tamil MN.ttc",
-            "\u{0BA4}\u{0BCA}\u{0B95}\u{0BC1}\u{0B95}\u{0BCD}\u{0B95}\u{0BAA}\u{0BCD}\u{0BAA}\
-             \u{0B9F}\u{0BCD}\u{0B9F}\u{0BC1}",
-            "--font-funcs ot",
-        ),
-        "tgm_e=0+1702|\
-         tgc_ta=0+1598|\
-         tgm_aa=0+1149|\
-         tgc_ku=2+1962|\
-         tgc_k=4+1592|\
-         tgc_ka=6+1592|\
-         tgc_p=7+1370|\
-         tgc_pa=9+1370|\
-         tgc_tt=10+1596|\
-         tgc_ttu=12+1833"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_092() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0041}\u{0066}\u{0300}\u{0066}\u{0069}\u{005A}",
-            "--font-funcs ot",
-        ),
-        "A=0+1479|\
-         f=1+682|\
-         gravecmb=1@-551,588+0|\
-         fi=3+1139|\
-         Z=5+1251"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_093() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{05E1}\u{05B0}",
-            "--font-funcs ot",
-        ),
-        "shevahebrew=0@51,0+0|\
-         samekhhebrew=0+1361"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_094() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0054}\u{0068}\u{0020}\u{0074}\u{0068}\u{0020}\u{006C}\u{006C}\u{0020}\u{0074}\
-             \u{0065}\u{0020}\u{0074}\u{006F}\u{0020}\u{0074}\u{0072}\u{0020}\u{0066}\u{0072}\
-             \u{0020}\u{0066}\u{0075}\u{0020}\u{0066}\u{006A}",
-            "--font-funcs ot",
-        ),
-        "T_h=0+2308|\
-         space=2+569|\
-         t_h=3+1687|\
-         space=5+569|\
-         l_l=6+1108|\
-         space=8+569|\
-         t_e=9+1408|\
-         space=11+569|\
-         t_o=12+1531|\
-         space=14+569|\
-         t_r=15+1385|\
-         space=17+569|\
-         f_r=18+1432|\
-         space=20+569|\
-         f_u=21+1733|\
-         space=23+569|\
-         f_j=24+1098"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_095() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0054}\u{0065}\u{0020}\u{0041}\u{0056}\u{0020}\u{0054}\u{0072}\u{0020}\u{0056}\
-             \u{0061}\u{0020}\u{0072}\u{0054}\u{0020}\u{0065}\u{0054}\u{0020}\u{0054}\u{0064}",
-            "--font-funcs ot",
-        ),
-        "T=0+1497|\
-         e=1@-62,0+699|\
-         space=2+569|\
-         A=3+1431|\
-         V=4@-37,0+1377|\
-         space=5+569|\
-         T=6+1510|\
-         r=7@-50,0+803|\
-         space=8+569|\
-         V=9+1376|\
-         a=10@-37,0+1014|\
-         space=11+569|\
-         r=12+853|\
-         T=13+1560|\
-         space=14+569|\
-         e=15+761|\
-         T=16+1560|\
-         space=17+569|\
-         T=18+1515|\
-         d=19@-45,0+1006"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_096() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0627}\u{0644}\u{0623}\u{064E}\u{0628}\u{0652}\u{062C}\u{064E}\u{062F}\u{0650}\
-             \u{064A}\u{064E}\u{0651}\u{0629}\u{0640}\u{0627}\u{0644}\u{0639}\u{064E}\u{0631}\
-             \u{064E}\u{0628}\u{0650}\u{064A}\u{064E}\u{0651}\u{0629}",
-            "--font-funcs ot",
-        ),
-        "u0629.final.tehMarbuta=26+713|\
-         u064e_u0651.shaddaFatha=23@0,-200+0|\
-         u064a.medial.yeh=23+656|\
-         u0650.kasra=21@80,290+80|\
-         u0628.initial.beh=21@-80,0+576|\
-         u064e.fatha=19@200,-570+200|\
-         u0631.final.reh=19@-200,0+702|\
-         u064e.fatha=17@200,-200+200|\
-         u0639.medial.ain=17@-200,0+738|\
-         u0644.initial.lam=16+515|\
-         u0627.final.alef=15+647|\
-         u0640.tatweel=14+449|\
-         u0629.final.tehMarbuta=13+713|\
-         u064e_u0651.shaddaFatha=10@0,-200+0|\
-         u064a.initial.yeh=10+656|\
-         u0650.kasra=8@80,570+80|\
-         u062f.final.dal=8@-80,0+822|\
-         u064e.fatha=6@290,-160+290|\
-         u062c.medial.jeem=6@-290,0+1069|\
-         u0652.sukun=4@0,-200+0|\
-         u0628.initial.beh=4+656|\
-         u064e.fatha=1@-372,120+-372|\
-         u0644_u0623.isolated.lamHamzaOnAlef=1@120,0+1282|\
-         u0627.alef=0+647"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_097() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0628}\u{064A}\u{064E}\u{0651}\u{0629}",
-            "--font-funcs ot",
-        ),
-        "u0629.final.tehMarbuta=4+713|\
-         u064e_u0651.shaddaFatha=1@0,-200+0|\
-         u064a.medial.yeh=1+656|\
-         u0628.initial.beh=0+656"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_098() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0631}\u{0628}",
-            "--font-funcs ot",
-        ),
-        "u0628.beh=1+1415|\
-         u0631.reh=0@-202,0+700"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_099() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0628}\u{064F}",
-            "--font-funcs ot",
-        ),
-        "u064f.damma=0@250,-250+250|\
-         u0628.beh=0@-250,0+1165"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_100() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0644}\u{064E}\u{0645}\u{064E}\u{0651}\u{0627}",
-            "--font-funcs ot",
-        ),
-        "u0627.final.alef=5+647|\
-         u064e.fatha=0@-80,160+-80|\
-         u064e_u0651.shaddaFatha=0@490,250+490|\
-         u0644_u0645.initial.lamMeem=0@-410,0+415"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_101() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Apple Color Emoji.ttc",
-            "\u{1F468}\u{200D}\u{1F469}\u{200D}\u{1F467}\u{200D}\u{1F466}",
-            "--remove-default-ignorables --font-funcs ot",
-        ),
-        "u1F9D1_u1F9D1_u1F9D2_u1F9D2=0+800"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_102() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Zapfino.ttf",
-            "\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\
-             \u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\
-             \u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\
-             \u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\
-             \u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{006F}",
-            "--font-funcs ot",
-        ),
-        "Z=0+416|\
-         a=1@-21,0+264|\
-         p_f=2+433|\
-         i=4+181|\
-         n=5+261|\
-         Z=6+416|\
-         a=7@-21,0+264|\
-         p_f=8+433|\
-         i=10+181|\
-         n=11+261|\
-         Z=12+416|\
-         a=13@-21,0+264|\
-         p_f=14+433|\
-         i=16+181|\
-         n=17+261|\
-         Z=18+416|\
-         a=19@-21,0+264|\
-         p_f=20+433|\
-         i=22+181|\
-         n=23+261|\
-         Z=24+416|\
-         a=25@-21,0+264|\
-         p_f=26+433|\
-         i=28+181|\
-         n=29+261|\
-         Z=30+416|\
-         a=31@-21,0+264|\
-         p_f=32+433|\
-         i=34+181|\
-         n=35+261|\
-         Z=36+416|\
-         a=37@-21,0+264|\
-         p_f=38+433|\
-         i=40+181|\
-         n=41+261|\
-         Z=42+416|\
-         a=43@-21,0+264|\
-         p_f=44+433|\
-         i=46+181|\
-         n=47+261|\
-         Z=48+416|\
-         a=49@-21,0+264|\
-         p_f=50+433|\
-         i=52+181|\
-         n=53+261|\
-         Z=54+416|\
-         a=55@-21,0+264|\
-         p_f=56+433|\
-         i=58+181|\
-         n=59+261|\
-         Z=60+416|\
-         a=61@-21,0+264|\
-         p_f=62+433|\
-         i=64+181|\
-         n=65+261|\
-         Z_a_p_f_i_n_o=66+2333"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_103() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Thonburi.ttc",
-            "\u{0E17}\u{0E35}\u{0E48}\u{0E4A}",
-            "--font-funcs ot",
-        ),
-        "thothahanthai_saraiithai_maiekthai=0+1616|\
-         maitrithai.key=0+1300"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_104() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0066}\u{0069}\u{006e}\u{0065}",
-            "--font-funcs ot",
-        ),
-        "fi=0+1139|\
-         n=2+1024|\
-         e=3+909"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_105() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0066}\u{0069}\u{006e}\u{0065}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "f=0+682|\
-         i=1+569|\
-         n=2+1024|\
-         e=3+909"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_106() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Kokonor.ttf",
-            "\u{0F62}\u{0F92}\u{0FB1}\u{0F74}",
-            "--font-funcs ot",
-        ),
-        "r_g_y_u=0+1579"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_107() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Kokonor.ttf",
-            "\u{0F62}\u{0F92}\u{0FB1}\u{0F74}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "r_g_y_u=0+1579"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_108() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_109() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features dlig=1",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_110() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_111() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features dlig=1,sups=0,tnum=0",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_112() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0,sups=0,tnum=0",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_113() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features smcp=1",
-        ),
-        "F.small=0+903|\
-         I.small=1+634|\
-         R.small=2+1113|\
-         S.small=3+911|\
-         T.small=4+1075"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_114() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0,dlig=1",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_117() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Helvetica.ttc",
-            "\u{006D}\u{0300}",
-            "--font-funcs ot",
-        ),
-        "m=0+1706|\
-         gravecmb=0@-284,10+0"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_118() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{006D}\u{0300}",
-            "--font-funcs ot",
-        ),
-        "mgrave=0+1912"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_119() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0066}\u{0069}",
-            "--font-funcs ot",
-        ),
-        "fi=0+1139"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_120() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Khmer MN.ttc",
-            "\u{17A2}\u{1780}\u{17D2}\u{179F}\u{179A}\u{1781}\u{17D2}\u{1798}\u{17C2}\u{179A}",
-            "--font-funcs ot",
-        ),
-        "km_qa=0+1230|\
-         km_ka=1+1230|\
-         km_sa.sub=1+620|\
-         km_ro=4+712|\
-         km_vs_ae=5+726|\
-         km_kha=5+1230|\
-         km_mo.sub=5+0|\
-         km_ro=9+712"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_121() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Tamil MN.ttc",
-            "\u{0BA4}\u{0BCA}\u{0B95}\u{0BC1}\u{0B95}\u{0BCD}\u{0B95}\u{0BAA}\u{0BCD}\u{0BAA}\
-             \u{0B9F}\u{0BCD}\u{0B9F}\u{0BC1}",
-            "--font-funcs ot",
-        ),
-        "tgm_e=0+1702|\
-         tgc_ta=0+1598|\
-         tgm_aa=0+1149|\
-         tgc_ku=2+1962|\
-         tgc_k=4+1592|\
-         tgc_ka=6+1592|\
-         tgc_p=7+1370|\
-         tgc_pa=9+1370|\
-         tgc_tt=10+1596|\
-         tgc_ttu=12+1833"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_122() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0041}\u{0066}\u{0300}\u{0066}\u{0069}\u{005A}",
-            "--font-funcs ot",
-        ),
-        "A=0+1479|\
-         f=1+682|\
-         gravecmb=1@-551,588+0|\
-         fi=3+1139|\
-         Z=5+1251"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_123() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{05E1}\u{05B0}",
-            "--font-funcs ot",
-        ),
-        "shevahebrew=0@51,0+0|\
-         samekhhebrew=0+1361"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_124() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0054}\u{0068}\u{0020}\u{0074}\u{0068}\u{0020}\u{006C}\u{006C}\u{0020}\u{0074}\
-             \u{0065}\u{0020}\u{0074}\u{006F}\u{0020}\u{0074}\u{0072}\u{0020}\u{0066}\u{0072}\
-             \u{0020}\u{0066}\u{0075}\u{0020}\u{0066}\u{006A}",
-            "--font-funcs ot",
-        ),
-        "T_h=0+2308|\
-         space=2+569|\
-         t_h=3+1687|\
-         space=5+569|\
-         l_l=6+1108|\
-         space=8+569|\
-         t_e=9+1408|\
-         space=11+569|\
-         t_o=12+1531|\
-         space=14+569|\
-         t_r=15+1385|\
-         space=17+569|\
-         f_r=18+1432|\
-         space=20+569|\
-         f_u=21+1733|\
-         space=23+569|\
-         f_j=24+1098"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_125() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0054}\u{0065}\u{0020}\u{0041}\u{0056}\u{0020}\u{0054}\u{0072}\u{0020}\u{0056}\
-             \u{0061}\u{0020}\u{0072}\u{0054}\u{0020}\u{0065}\u{0054}\u{0020}\u{0054}\u{0064}",
-            "--font-funcs ot",
-        ),
-        "T=0+1497|\
-         e=1@-62,0+699|\
-         space=2+569|\
-         A=3+1431|\
-         V=4@-37,0+1377|\
-         space=5+569|\
-         T=6+1510|\
-         r=7@-50,0+803|\
-         space=8+569|\
-         V=9+1376|\
-         a=10@-37,0+1014|\
-         space=11+569|\
-         r=12+853|\
-         T=13+1560|\
-         space=14+569|\
-         e=15+761|\
-         T=16+1560|\
-         space=17+569|\
-         T=18+1515|\
-         d=19@-45,0+1006"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_126() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0627}\u{0644}\u{0623}\u{064E}\u{0628}\u{0652}\u{062C}\u{064E}\u{062F}\u{0650}\
-             \u{064A}\u{064E}\u{0651}\u{0629}\u{0640}\u{0627}\u{0644}\u{0639}\u{064E}\u{0631}\
-             \u{064E}\u{0628}\u{0650}\u{064A}\u{064E}\u{0651}\u{0629}",
-            "--font-funcs ot",
-        ),
-        "u0629.final.tehMarbuta=26+713|\
-         u064e_u0651.shaddaFatha=23@0,-200+0|\
-         u064a.medial.yeh=23+656|\
-         u0650.kasra=21@80,290+80|\
-         u0628.initial.beh=21@-80,0+576|\
-         u064e.fatha=19@200,-570+200|\
-         u0631.final.reh=19@-200,0+702|\
-         u064e.fatha=17@200,-200+200|\
-         u0639.medial.ain=17@-200,0+738|\
-         u0644.initial.lam=16+515|\
-         u0627.final.alef=15+647|\
-         u0640.tatweel=14+449|\
-         u0629.final.tehMarbuta=13+713|\
-         u064e_u0651.shaddaFatha=10@0,-200+0|\
-         u064a.initial.yeh=10+656|\
-         u0650.kasra=8@80,570+80|\
-         u062f.final.dal=8@-80,0+822|\
-         u064e.fatha=6@290,-160+290|\
-         u062c.medial.jeem=6@-290,0+1069|\
-         u0652.sukun=4@0,-200+0|\
-         u0628.initial.beh=4+656|\
-         u064e.fatha=1@-372,120+-372|\
-         u0644_u0623.isolated.lamHamzaOnAlef=1@120,0+1282|\
-         u0627.alef=0+647"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_127() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0628}\u{064A}\u{064E}\u{0651}\u{0629}",
-            "--font-funcs ot",
-        ),
-        "u0629.final.tehMarbuta=4+713|\
-         u064e_u0651.shaddaFatha=1@0,-200+0|\
-         u064a.medial.yeh=1+656|\
-         u0628.initial.beh=0+656"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_128() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0631}\u{0628}",
-            "--font-funcs ot",
-        ),
-        "u0628.beh=1+1415|\
-         u0631.reh=0@-202,0+700"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_129() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0628}\u{064F}",
-            "--font-funcs ot",
-        ),
-        "u064f.damma=0@250,-250+250|\
-         u0628.beh=0@-250,0+1165"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_130() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0644}\u{064E}\u{0645}\u{064E}\u{0651}\u{0627}",
-            "--font-funcs ot",
-        ),
-        "u0627.final.alef=5+647|\
-         u064e.fatha=0@-80,160+-80|\
-         u064e_u0651.shaddaFatha=0@490,250+490|\
-         u0644_u0645.initial.lamMeem=0@-410,0+415"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_131() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0644}\u{0627}",
-            "--font-funcs ot",
-        ),
-        "u0644_u0627.isolated.lamAlef=0+1162"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_132() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/GeezaPro.ttc",
-            "\u{0644}\u{0627}",
-            "--font-funcs ot --features rlig=0",
-        ),
-        "u0627.final.alef=1+647|\
-         u0644.initial.lam=0+515"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_134() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/SFNS.ttf",
-            "\u{0054}\u{0065}\u{0020}\u{0041}\u{0056}\u{0020}\u{0054}\u{0072}\u{0020}\u{0056}\
-             \u{0061}\u{0020}\u{0072}\u{0054}\u{0020}\u{0065}\u{0054}\u{0020}\u{0054}\u{0064}",
-            "--font-funcs ot",
-        ),
-        "T=0+920|\
-         e=1+1049|\
-         space=2+420|\
-         A=3+1141|\
-         V=4+1292|\
-         space=5+420|\
-         T=6+960|\
-         r=7+631|\
-         space=8+420|\
-         V=9+1142|\
-         a=10+1028|\
-         space=11+420|\
-         r=12+461|\
-         T=13+1190|\
-         space=14+420|\
-         e=15+779|\
-         T=16+1190|\
-         space=17+420|\
-         T=18+920|\
-         d=19+1134"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_135() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/SFNS.ttf",
-            "\u{0054}\u{0065}\u{0020}\u{0041}\u{0056}\u{0020}\u{0054}\u{0072}\u{0020}\u{0056}\
-             \u{0061}\u{0020}\u{0072}\u{0054}\u{0020}\u{0065}\u{0054}\u{0020}\u{0054}\u{0064}",
-            "--font-ptem 9 --font-funcs ot",
-        ),
-        "T=0@19,0+958|\
-         e=1@19,0+1087|\
-         space=2@19,0+458|\
-         A=3@19,0+1179|\
-         V=4@19,0+1330|\
-         space=5@19,0+458|\
-         T=6@19,0+998|\
-         r=7@19,0+669|\
-         space=8@19,0+458|\
-         V=9@19,0+1180|\
-         a=10@19,0+1066|\
-         space=11@19,0+458|\
-         r=12@19,0+499|\
-         T=13@19,0+1228|\
-         space=14@19,0+458|\
-         e=15@19,0+817|\
-         T=16@19,0+1228|\
-         space=17@19,0+458|\
-         T=18@19,0+958|\
-         d=19@19,0+1172"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_136() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Apple Color Emoji.ttc",
-            "\u{1F468}\u{200D}\u{1F469}\u{200D}\u{1F467}\u{200D}\u{1F466}",
-            "--remove-default-ignorables --font-funcs ot",
-        ),
-        "u1F9D1_u1F9D1_u1F9D2_u1F9D2=0+800"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_137() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Zapfino.ttf",
-            "\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\
-             \u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\
-             \u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\
-             \u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\
-             \u{005A}\u{0061}\u{0070}\u{0066}\u{0069}\u{006E}\u{005A}\u{0061}\u{0070}\u{0066}\
-             \u{0069}\u{006E}\u{006F}",
-            "--font-funcs ot",
-        ),
-        "Z=0+416|\
-         a=1@-21,0+264|\
-         p_f=2+433|\
-         i=4+181|\
-         n=5+261|\
-         Z=6+416|\
-         a=7@-21,0+264|\
-         p_f=8+433|\
-         i=10+181|\
-         n=11+261|\
-         Z=12+416|\
-         a=13@-21,0+264|\
-         p_f=14+433|\
-         i=16+181|\
-         n=17+261|\
-         Z=18+416|\
-         a=19@-21,0+264|\
-         p_f=20+433|\
-         i=22+181|\
-         n=23+261|\
-         Z=24+416|\
-         a=25@-21,0+264|\
-         p_f=26+433|\
-         i=28+181|\
-         n=29+261|\
-         Z=30+416|\
-         a=31@-21,0+264|\
-         p_f=32+433|\
-         i=34+181|\
-         n=35+261|\
-         Z=36+416|\
-         a=37@-21,0+264|\
-         p_f=38+433|\
-         i=40+181|\
-         n=41+261|\
-         Z=42+416|\
-         a=43@-21,0+264|\
-         p_f=44+433|\
-         i=46+181|\
-         n=47+261|\
-         Z=48+416|\
-         a=49@-21,0+264|\
-         p_f=50+433|\
-         i=52+181|\
-         n=53+261|\
-         Z=54+416|\
-         a=55@-21,0+264|\
-         p_f=56+433|\
-         i=58+181|\
-         n=59+261|\
-         Z=60+416|\
-         a=61@-21,0+264|\
-         p_f=62+433|\
-         i=64+181|\
-         n=65+261|\
-         Z_a_p_f_i_n_o=66+2333"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_138() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Thonburi.ttc",
-            "\u{0E17}\u{0E35}\u{0E48}\u{0E4A}",
-            "--font-funcs ot",
-        ),
-        "thothahanthai_saraiithai_maiekthai=0+1616|\
-         maitrithai.key=0+1300"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_139() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0066}\u{0069}\u{006e}\u{0065}",
-            "--font-funcs ot",
-        ),
-        "fi=0+1139|\
-         n=2+1024|\
-         e=3+909"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_140() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Times.ttc",
-            "\u{0066}\u{0069}\u{006e}\u{0065}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "f=0+682|\
-         i=1+569|\
-         n=2+1024|\
-         e=3+909"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_141() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Kokonor.ttf",
-            "\u{0F62}\u{0F92}\u{0FB1}\u{0F74}",
-            "--font-funcs ot",
-        ),
-        "r_g_y_u=0+1579"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_142() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Kokonor.ttf",
-            "\u{0F62}\u{0F92}\u{0FB1}\u{0F74}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "r_g_y_u=0+1579"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_143() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_144() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features dlig=1",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_145() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_146() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features dlig=1,sups=0,tnum=0",
-        ),
-        "f_i=0+1097|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_147() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0,sups=0,tnum=0",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s=3+728|\
-         t=4+725"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_148() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features smcp=1",
-        ),
-        "F.small=0+903|\
-         I.small=1+634|\
-         R.small=2+1113|\
-         S.small=3+911|\
-         T.small=4+1075"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_149() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Apple Chancery.ttf",
-            "\u{0066}\u{0069}\u{0072}\u{0073}\u{0074}",
-            "--font-funcs ot --features liga=0,dlig=1",
-        ),
-        "f=0+639|\
-         i=1+606|\
-         r=2+853|\
-         s_t=3+1438"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_150() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Skia.ttf",
-            "\u{0041}\u{0056}",
-            "--font-funcs ot",
-        ),
-        "A=0+1345|\
-         V=1@-12,0+1346"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_151() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Supplemental/Devanagari Sangam MN.ttc",
-            "\u{092D}\u{0941}",
-            "--font-funcs ot",
-        ),
-        "dn_bha=0+1339|\
-         dn_u_matra.mrk=0@-296,11+0"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_154() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Apple Color Emoji.ttc",
-            "\u{1F469}\u{1F3FD}\u{200D}\u{1F91D}\u{200D}\u{1F468}\u{1F3FE}",
-            "--font-funcs ot",
-        ),
-        "u1F469.3.L=0+800|\
-         u1F468.4.RA=0@-800,0+0"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_157() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/Apple Color Emoji.ttc",
-            "\u{1FAF1}\u{1F3FF}\u{200D}\u{1FAF2}\u{1F3FC}",
-            "--font-funcs ot",
-        ),
-        "u1FAF1.5.L=0+800|\
-         u1FAF2.2.R=0@-800,0+0"
-    );
-}
-
-#[cfg(target_os = "macos")]
-#[test]
-fn macos_160() {
-    assert_eq!(
-        shape(
-            "/System/Library/Fonts/LucidaGrande.ttc",
-            "\u{0066}\u{0066}\u{0020}\u{0066}\u{0066}\u{0066}\u{0020}\u{0066}\u{0066}",
-            "--features=-liga[3:5]",
-        ),
-        "ff=0+1443|\
-         space=2+648|\
-         f=3+753|\
-         f=4+753|\
-         f=5+753|\
-         space=6+648|\
-         ff=7+1443"
-    );
-}
-
 #[test]
 fn mark_attachment_001() {
     assert_eq!(
