+++ tests/shaping/main.rs
@@ -1,6 +1,7 @@
 mod aots;
 mod extra;
 mod in_house;
+mod macos;
 mod text_rendering_tests;
 
 use std::str::FromStr;
